+++ src/_dep.rs
@@ -30,7 +30,7 @@ pub extern crate alloc as _alloc;
 #[doc(inline)]
 pub use ::std as _std;
 
-/* 40 optional dependencies */
+/* 42 optional dependencies */
 // In sync with ../Cargo.toml::dep_all & ../config/dep_all.rs
 
 reexport! { optional_crate (unsafe) "dep_allocator_api2", "allocator-api2", allocator_api2,
@@ -104,7 +104,10 @@ reexport! { optional_crate (unsafe) "dep_raw_cpuid", "raw-cpuid", raw_cpuid,
     doc: "A library to parse the x86 CPUID instruction."
 }
 reexport! { optional_crate (safe) "dep_regex_lite", "regex-lite", regex_lite,
-    doc: " A lightweight regex engine optimized for binary size and compilation time."
+    doc: "A lightweight regex engine optimized for binary size and compilation time."
+}
+reexport! { optional_crate (safe) "dep_ring", "ring", ring,
+    doc: "Safe, fast, small crypto using Rust."
 }
 reexport! { optional_crate (unsafe) "dep_rkyv", "rkyv", rkyv,
     doc: "A zero-copy deserialization framework for Rust."
@@ -148,6 +151,9 @@ reexport! { optional_crate (safe)
 }
 reexport! { optional_crate (safe) "dep_unicode_width", "unicode-width", unicode_width,
     doc: "Determine displayed width of `char` and `str` types."
+}
+reexport! { optional_crate (safe) "dep_ureq", "ureq", ureq,
+    doc: "Simple, safe HTTP client."
 }
 reexport! { optional_crate (unsafe) "dep_wasm_bindgen", "wasm-bindgen", wasm_bindgen,
     doc: "Easy support for interacting between JS and Rust."