+++ DOCS/CHANGELOG.md
@@ -19,7 +19,7 @@
   - `ExtFuture`: `pending`, `poll_fn`, `ready`.
   - prngs: `from_state`, `inner_state`.
 - new re-exports: `SystemAlloc`, `std::{env::*, process::*}`, `::log::*`
-- new optional dependencies: `itertools`.
+- new optional dependencies: `itertools`, `ring`, `ureq`.
 - add musl architectures to `check.rs` script.
 
 ### Removed